pyproject.toml
[project]
name = "pr-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-search-documents>=11.6.0",
    "openai>=2.6.1",
    "pygithub>=2.8.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
